Add a list of FoodProperties to Food class
Reason: Previously Food used to have a variable of an appropriate type
for each property a food can have. With this change, it will instead
maintain a list of FoodProperties that can automatically be created and
dynamically be adjusted using the data from the database. Each column in
the ingredients-table in the database will be represented by a
FoodProperty.

Signed-off-by: Chris <uni@hermannc.de>